+++ docs/source/index.md
@@ -26,6 +26,7 @@ Below is the current list of TRL trainers, organized by method type (⚡️ = vL
 - [`RLOOTrainer`] ⚡️
 - [`OnlineDPOTrainer`] ⚡️
 - [`NashMDTrainer`] ⚡️
+- [`experimental.xpo.XPOTrainer`] 🧪 ⚡️
 - [`PPOTrainer`]
 
 ### Reward modeling
@@ -44,10 +45,6 @@ Below is the current list of TRL trainers, organized by method type (⚡️ = vL
 - [`experimental.bco.BCOTrainer`] 🧪
 - [`CPOTrainer`]
 - [`KTOTrainer`]
-
-### Experimental
-
-- [`experimental.xpo.XPOTrainer`] 🧪 ⚡️
 
 ### Knowledge distillation
