+++ trl/experimental/xpo/xpo_trainer.py
@@ -69,7 +69,7 @@ class XPOTrainer(OnlineDPOTrainer):
             [`~transformers.AutoModelForSequenceClassification`].
         judge ([`BasePairwiseJudge`]):
             The judge to use for pairwise comparison of model completions.
-        args ([`XPOConfig`]):
+        args ([`experimental.xpo.XPOConfig`]):
             The XPO config arguments to use for training.
         data_collator ([`~transformers.DataCollator`]):
             The data collator to use for training. If None is specified, the default data collator
